requirements.txt
django>=5.2.1
djangorestframework>=3.16.0
nltk>=3.8
sentence-transformers>=4.1.0